part1_implementation/requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Hugging Face + NLP
datasets>=2.19.0
transformers>=4.40.0
huggingface_hub>=0.23.0
accelerate>=0.30.0

# Torch (CPU is fine)
torch

# Metrics + plots
scikit-learn>=1.3.0
matplotlib>=3.7.0
tensorboard